+++ cli/src/lib.rs
@@ -1,6 +1,5 @@
 pub mod meta;
 pub mod solc;
-pub mod utils;
 pub mod subgraph;
 
 // re-export main types and functionalities
