+++ cli/src/main.rs
@@ -1,7 +1,6 @@
 pub mod meta;
 pub(crate) mod cli;
 pub(crate) mod solc;
-pub(crate) mod utils;
 pub(crate) mod subgraph;
 
 #[tokio::main]
